02_productivity/shiny_app/requirements.txt
shiny>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=1.5.0
